// src/api/generated/model/createCreativeTaskRequestDtoTargetPlatform.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Целевая платформа публикации
 */
export type CreateCreativeTaskRequestDtoTargetPlatform = typeof CreateCreativeTaskRequestDtoTargetPlatform[keyof typeof CreateCreativeTaskRequestDtoTargetPlatform];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateCreativeTaskRequestDtoTargetPlatform = {
  VK_GROUP: 'VK_GROUP',
  VK_USER: 'VK_USER',
  YOUTUBE_CHANNEL: 'YOUTUBE_CHANNEL',
  RUTUBE_CHANNEL: 'RUTUBE_CHANNEL',
  TG_CHANNEL: 'TG_CHANNEL',
  TG_BOT: 'TG_BOT',
  MAX_BOT: 'MAX_BOT',
  MAX_CHANNEL: 'MAX_CHANNEL',
} as const;